bugfix post reaction pipeline

+++ Service/PostReactionService.cs
@@ -1,5 +1,6 @@
 ﻿using DAL.Repositories.Abstract;
 using Domain;
+using Domain.Exceptions;
 using Domain.Models.Pagination;
 using Service.Abstract;
 
@@ -8,15 +9,12 @@ namespace Service
     public class PostReactionService : IPostReactionService
     {
         private readonly IPostReactionRepository _postReactionRepository;
+        private readonly IPostRepository _postRepository;
 
-        public PostReactionService(IPostReactionRepository postReactionRepository)
+        public PostReactionService(IPostReactionRepository postReactionRepository, IPostRepository postRepository)
         {
             _postReactionRepository = postReactionRepository;
-        }
-
-        public async Task Add(PostReaction entity)
-        {
-            await _postReactionRepository.AddAsync(entity);
+            _postRepository = postRepository;
         }
 
         public async Task<IEnumerable<PostReaction>> GetAll()
@@ -38,19 +36,68 @@ namespace Service
         {
             return await _postReactionRepository.GetPagedData(query);
         }
+        
+        public async Task Add(PostReaction entity)
+        {
+            if (await PostDoesNotExistAsync(entity.PostId))
+            {
+                throw new ValidationException($"{nameof(Post)} of ID: {entity.PostId} does not exist");
+            }
+
+            if (await ExistsSuchReactionAsync(entity.PostId,entity.UserId))
+            {
+                throw new ValidationException($"{nameof(Post)}ID: {entity.PostId} already has found from {nameof(User)}ID: {entity.UserId}");
+            }
+
+            await _postReactionRepository.AddAsync(entity);
+            await _postReactionRepository.SaveChangesAsync();
+        }
 
         public async Task<bool> Remove(int id, int issuerId)
         {
-            await _postReactionRepository.RemoveAsync(id);
+            var reaction = await _postReactionRepository.GetByIdAsync(id);
 
+            if (reaction.UserId != issuerId)
+            {
+                throw new ValidationException($"This {nameof(PostReaction)} does not belong to authorized user");
+            }
+
+            await _postReactionRepository.RemoveAsync(id);
+            await _postReactionRepository.SaveChangesAsync();
             return true;
         }
 
         public async Task<bool> Update(PostReaction entity)
         {
-            await _postReactionRepository.UpdateAsync(entity);
+            var found = await _postReactionRepository.GetWhereAsync(r => r.PostId == entity.PostId && r.UserId == entity.UserId);
+
+            var post = found.FirstOrDefault();
+
+            if (post == null)
+            {
+                throw new ValidationException($"{nameof(PostReaction)} of ID: {entity.PostId} does not exist");
+            }
+
+            post.ReactionType = entity.ReactionType;
+
+            await _postReactionRepository.UpdateAsync(post);
+            await _postReactionRepository.SaveChangesAsync();
 
             return true;
+        }
+
+        private async Task<bool> ExistsSuchReactionAsync(int postId, int userId)
+        {
+            var requestedPostReactionsPostedByUser = await _postReactionRepository.GetWhereAsync(r => r.PostId == postId && r.UserId == userId);
+
+            return requestedPostReactionsPostedByUser.Any();
+        }
+
+        private async Task<bool> PostDoesNotExistAsync(int postId)
+        {
+            var searchRequestForPost = await _postRepository.GetWhereAsync(e => e.Id == postId);
+
+            return !searchRequestForPost.Any();
         }
     }
 }
